Make exception message argument optional
The error_message attribute of exceptions is still a str. All exceptions
are expected to set the error_message to a value that makes sense for
the exception, something human readable explaining the exception.

An exception can also allow the message to be overridden by the user.

They argument order of KeyError is now reversed. We always have to
provide a key but the message is optional.