Print build progress / result
When running acton test we implicitly build the project to ensure all
tests are up to date before actually running the tests. actonc was
invoked with --quiet and we wouldn't print any progress which meant for
larger projects it felt like the program just hung. Now we print the
normal build output in real time.

Similarly if there is an compilation error the output from the compiler
is printed so it's clear what the problem is.