Accept larger time diff in test_time
Bump up the accepted diff. It has started failing quite a lot in CI
lately, with the diff being like 7 seconds. I think it's because of how
we run the program from test.hs. We first get the time then build and
run the test program. Presumably the build takes a little time and that
leads to the larger diff. Maybe our builds are slightly slower
now (after linking libc++??) or the CI runners are slightly slower.

If there's an actual problem with the time function it's likely to be
off by much more.