Revamp testing
Tests are now run multiple times for a minimum of
1ms (MIN_TEST_DURATION). Many small unit tests will run a bunch of times
and we can thus get an average for performance numbers and we can detect
flakiness. The whole display routine has been rewritten to be able to
show such results. It also shows tests in progress in a more consistent
manner.

Errors / failures were flipped. Now fixed.

Async actor test and env tests are now run properly. Previously we would
actually start all such async tests at once, so they would run in
parallel. The idea is that we should grab one test, let it run and once
complete we move on to the next. This has now been corrected so it works
as intended.

There's a bunch of to_json / from_json serialization functions but it
doesn't work very well for some reason so I've had to go with separate
more manual functions for now. Should look into this.

Split TestDisplay class to instead be one for the module test and one
for the project test that assembles a summary view of tests in all
modules. Makes for simpler code.

Avoid named tuples for keeping the test information since we have some
type inference bugs. I wish I could find smaller reproduction cases for
this but haven't had time. Using classes is more stable, thus TestInfo!