Revamp stdin_install & add env actions
There are now new methods on env to interact with environment variables:
- getenv
- setenv
- unsetenv
- getenvb
- setenvb
- unsetenvb

The first three work with str input and output, presuming data is
encoded with UTF-8. getenvb and friends instead use bytes as input and
output for raw access to environment variables.

Some time ago stdin_install was changed from working with str to be
bytes instead, which is arguably more correct. However, a bug crept in
so while the returned data was indeed bytes, the type signature said
str. This has been fixed by revamping stdin_install. stdin_install now
support working with strings in a convenient manner just like
previously, but now handles chunking of data as well as reporting errors
(which were previously unhandled) as well as providing a raw bytes
interface.

When stdin is attached to a terminal, it uses line buffering. The buffer
is however limited in size so if the line is really long, it will be
chopped into chunks. If a unicode characters happens to be split so we
only get the first byte of a multi-byte character, then the decode would
fail. The string StringDecoder actor will attempt to handle this and
store the remaining bytes in a buffer to be decoded upon receipt of the
next chunk. Just like with a raw interface to stdin, it attempts to be
line buffered but there is no guarantee when the line exceeds the buffer
size.

We attempt to automatically detect the encoding of the terminal by
reading the LANG environment variable. If this is not possible, we
assume UTF-8 encoding. Since we do not currently support anything
besides UTF-8, any other encoding will lead to an exception (which might
go unhandled).

Acton-by-Example has been updated with some examples around this.