Reduce variables sent to builder
Instead of sending all paths, we send the few paths that actually need
to be sent and compute the other ones based on relative paths in builder.

projpath is discovered by looking at current working directory since
builder is invoked in a project folder.

syspath is sent from actonc but all other variables are computed based
on these.

The downside right now is that the path separation character / is
hard-coded in builder which probably does not work on Windows. We'll
have to address that separately.