Improve project dependencies
actonc now takes an explicit --deppath argument which helps us with
nested dependencies, e.g. our top level project depends on A which in
turn depends on B. We want to place both A & B in the top/deps directory
and not deps/A and then deps/A/deps/B. Since the deps_path was
previously relative to the project dir, it wouldn't find project
dependencies that are sibling of the own project. It will now!

If --deppath is not specified we still use the relative deps, which
makes running actonc build (not acton build) still a viable alternative
for just finding .ty files etc in deps.