Reorder process args & add RunProcess
Move workdir & env to end of list and make them optional. Nicer interface!

Add an optional timeout argument so we can say we want to terminate a
program after some timeout period.

Added a new RunProcess which buffers stdout & stderr while waiting for
the program to exit after which a callback is called. This is a much
simpler interface for the scenario when we want to await a process to
finish and then read the output.