Add __cleanup__ to actors
Install GC finalizer __cleanup__ action for actors.

Add a new GC finalizer function per actor class and install it at actor
creation. The finalizer function in turn calls the __cleanup__ method on
the actor which can be used to clean things up. This is intended to be
used in I/O actors primarily.

We only install the cleanup action if it is defined for an actor.