Run arm64 on arm64
I just added macos-14 to the list so we would test another version of
macos, but the GitHub macos-14 runner is powered by Apple M1 and not
x86_64, so we don't just get another version but another CPU as well.
Our run-macos job grabs the acton version built by the test-macos-12
job, which we partially use to check compatibility of our compiled
artifacts across a range of macos versions. Interestingly it just worked
out of the box for macos-14 despite macos-14 being an arm64 and the
acton binary being built on x86_64. Cool stuff! I guess it does Rosetta
translation out of the box.

Regardless, we primarily want to test arm64 with arm64, so we're
switching it up here. The run-macos job for macos-14 will grab the
artifact from macos-14. Other run-macos jobs will continue grabbing the
macos-12 artifact.