Assert exception handling & wctx->id
Have seen some segfaults that seem to come from access jump0->exval and
presumably jump0 is not set... this was likely related to GC and TLS
which is addressed by using WorkerCtx, but nonetheless it's nice with
some asserts.

I've seen some rather rare segfaults in DEQ_ready where the worker
thread id is something like 191255, so way off our number of worker
threads. Asserting here in hopes of catching it.