Stop redirecting malloc & friends
Since the introduction of the GC, we have been making use of link time
redirection of malloc and friends in order to catch all invocations of
these calls. It works but is less flexible as we are forced to use the
GC-heap for everything. We would like to be able to mix GC-managed
memory with manual memory management and thus we want to avoid doing
link time redirection of malloc.

For a period of time, we've used a sort of hack where we dig out the
real malloc through dlsym shenanigans, but that only works when malloc
is in a libc that is dynamically linked. It won't work when statically
linking libc. We do this so we can run with manual malloc during module
init and avoid placing module constants on the GC-heap.

To disable link time redirection, we must make sure that all mallocs
that we want to be on the GC-heap must be explicit calls to GC_malloc.
Some time ago I patched all our external C library dependencies to
support pluggable malloc and we thus support this.

This commit now disables link time redirection. I have replaced all
remaining calls to malloc & friends with either acton_malloc() or
GC_malloc for cases when we always want to use the GC-heap. It should
now be possible to do manual memory management selectively, although we
don't have any real uses cases for it just yet.