Fix GC deadlock & uv_spawn
libuv blocks most signals when it forks in the uv_spawn() call. I guess
because signal handling can get messy around a fork.

On Unix like systems, the GC uses SIGPWR & SIGXCPU to pause & resume
threads. We must not ignore those signals since it leads to a deadlock.
When one thread initiates GC stop-the-world it will wait for all other
threads to comply and if they ignore signals, they will never comply.

Bump version of libuv to one which is patched to not block SIGPWR &
SIGXCPU!

Also add a test program that runs lots of subprocesses so we get to
exercise this code. I can reproduce the bug on like the 4th process down
and we run 1000 so I'm quite confident in this fix working well.