Improve crash handler with appname & gdb
appname is now included in the crash handlers so when we get a crash we
can easily distinguish which program produced it. This is helpful since
sometimes we are running subprocesses, as in the case with `acton test`.

Turn off confirmation, paginations etc for gdb which improves the
overall experience.