Silence compiler warnings
I don't quite get how this helps but it's some form of more explicit
cast. I mean, it's safe since when we deserialize we put those small
ints in the first place, so data can safely be truncated to the smaller
type, we just need the type system to agree and not complain.