Support term settings, canonical & echo
The canonical mode is the normal mode for a terminal where input is line
buffered and the buffer is managed by something else, so there might be
editing capabilities. Only when enter is hit, the line is sent to us. If
we want to react to individual characters, we must turn this off, which
is now possible by doing `env.set_stdin(canonical=False)`.

It is also possible to turn off the terminal echo with
`env.set_stdin(echo=False)`.

On startup, we read the terminal settings and store so that on shutdown
we restore the terminal settings.