Keep track of mem stats for tests
We look at the memory usage on the GC heap and the RSS of the program.
While there is a gc memory usage function, it shows the currently used
memory so we can't pick up memory that was both allocated and freed by
the GC during a test run, so instead we read the total number of bytes
allocated by the GC which should cover the memory usage even if the GC
has run.

RSS is collected in an attempt to catch memory leaks on the non-GC heap,
i.e. using malloc and not GC_MALLOC. It's probably not super reliable,
but at least it's unlikely to have few false positives when looking
across a large set of test results. For example, if we run 1000 runs of
the same test function, the memory usage might increase for the first
iteration but should then be relatively stable. If we have a memory leak
on the other hand, we are likely to see a memory increase on the
majority of test runs.