Place jump0 in wctx
We just want to initialize jump0 once per worker thread which we now
achieve by placing jump0 in our worker context wctx. Previous to this,
we would rerun the $PUSH every time we entered wt_work_cb, and thus
actually push a new exception jump point to the stack - effectively a
memory "leak" with run-away memory consumption as a consequence.