Refactor testing, reduce to single test executor
After having sorted out the test discovery to create proper objects
based on the kind of test we can now rely on that to separate tests
which in turn enables us to collapse the four separate testing pipelines
into a single one capable of running all kinds of tests! Sync tests are
treated as async and thus we can look at all tests as async.