Increase randomness
Rather than seed our PRNG with time() which is second-granularity, we
now feed in nanoseconds.

Running the test suite over and over would result in flakiness as we'd
get the same random values and that in turn affects some of the file
tests where it creates temp files (which would be the same due to random
being the same). This fixes that.