Place wctxs in global array to avoid GC
I figured placing wctx on the stack in the main_loop would fix our
issues with the GC cleaning up the WorkerCtx, but alas, the compiler
aggresively optimizes away the wctx since it's "not used much", like we
just do a pthread_setspecific etc so it ends up in some temporary
register and later on its not visible anymore and thus cleaned up.

I've run the test thousands of times now, there are some other
intermittent hangs, but no segfault crashes. Before this patch, I would
get segfaults probably on the order of 1 in 30-50, so maybe 2-3% and
that seems entirely gone now.