Add to_str_noc(opy)
to_str_noc will create a new Acton str (B_str) from a C string without
any form of processing. Useful when we have an already known-good C
string.

... and use it in int->str conversion to get rid of one copy.

It is probably not Unicode safe as-is since nchars is set to nbytes
which is only correct for ASCII / single-byte characters.