Replace asprintf with $FORMAT
Now that we have a slick $FORMAT that uses the right malloc and doesn't
unnecessarily copy data around etc we can use it for doing string
formatting like everywhere we expect an Acton string as return value.
Even when we don't we can format like this and just do fromB_str() on it
to get a C string.