Add perf test to CI tests
This compares performance between acton tip produced by the main branch
and the current branch. We use normal GitHub Action runners but as we
run both perf tests on the same computer, it has at least somewhat of a
chance to be reasonably accurate (-CPU throttling / congestion over time)

We're currently running this on the normal stdlib tests since those are
just some of the better tests we have. Not in any way comprehensive but
better than nothing.