Reduce PR CI jobs & do nightly
In order to get decent speed on CI builds we are using caching. Github
limits the cache size to 10GB in total and with all the jobs we have we
reach above that. MacOS 14 was the last one we added which tipped us
over 10GB so Github evicts one of the caches but since we're really
using all of them we just end up with lots of churn. I tried to use a
common cache among similar distros, like multiple versions of Debian, to
increase reuse but somewhat unsurprisingly, it doesn't work out great.

Ever since we started vendoring in libraries and building it ourselves,
the problems with supporting multiple versions of distributions and OS
has been at a minimum. There have been very few issues, so arguably
running all these tests for every PR is overkill. Thus, we're now
reducing the PR builds to much fewer OSes which makes it possible to use
caches for those and achieve fast PR builds.

There is a new daily job which runs that runs all the distro versions,
but without a cache, so we still get testing coverage albeit perhaps
somewhat delayed. This is very likely going to work out well in reality.