Reduce PR CI jobs & do nightly
In order to get decent speed on CI builds we are using caching. Github
limits the cache size to 10GB in total and with all the jobs we have we
reach above that. MacOS 14 was the last one we added which tipped us
over 10GB so Github evicts one of the caches but since we're really
using all of them we just end up with lots of churn. I tried to use a
common cache among similar distros, like multiple versions of Debian, to
increase reuse but somewhat unsurprisingly, it doesn't work out great.

Ever since we started vendoring in libraries and building it ourselves,
the challenges with supporting multiple versions of distributions and OS
has been at a minimum. There have been very few issues, so arguably
running all these tests for every PR is overkill. Thus, we're now
reducing the PR builds to much fewer OSes which makes it possible to use
caches for those and achieve fast PR builds. Currently, that includes:
- Linux x86_64 based on Debian 12
- MacOS 12 on x86_64
- MacOS 14 on aarch64

Using MacOS 12 is a bit of a compromise. The MacOS 14 runners do not
just imply a particular version of macos but they also all run on
aarch64 (M1). They are the fastest Macos runners and covering aarch64 is
a good thing. Older MacOS runners are x86_64. MacOS 13 has 4 CPUs and
tends to be faster than macos 12 which only has 3 (and perhaps slower
cores too?), so MacOS 13 is a good choice and so if we are to keep down
the number of jobs, macos 13 on x86_64 and macos 14 on aarch64 would
seem like a good pick. However, we use the oldest among x86_64 platforms
for the run-macos and upload build artifacts for releases, so from that
perspective it's good using macos-12 and that is why our PR builds use
macos-12 and macos-14.

There is a new daily job which runs that runs all the distro versions,
but without a cache, so we still get testing coverage albeit at a
nightly cadence rather than before-merging-the-PR. I think this will
work well in practice.

The GitHub Action CI config is now slightly convoluted, which I would
describe as a limitation of GH Action config. We now need an extra job
ahead of the normal test-job that first constructs a matrix for what
test jobs to run. Since we have separate test jobs for linux and macos,
we need two new matrix-making jobs. Sigh, but it works...