Support no-threads RTS mode
This adds a mode for running the RTS without threads. It can be useful
on platforms that do not have POSIX thread available, such as
WASM (although we are far from being able to run on WASM for other
reasons).

Since one of the goals is to remove the needs for pthreads, this must be
chosen at compile time rather than run time. Use the --no-threads flag:

  actonc --no-threads foo.act

The DB support currently requires threads. To get past compilation,
pthread is made optional and instead results in an error if attempted to
be used at run time.