Remove RTSmon
This removes the support for mon socket, mon log and mon on exit. The
idea is that we should instead expose RTS monitoring data via other
means. The underlying data that we exposed via RTSmon is now available
in the acton.rts module from:
- rts_stats(): statistics per worker thread
- db_actors(): list of all actors (when running with DB)
- db_servers(): list of DB servers in cluster
- db_rtses(): list of RTS nodes in cluster

These functions are the minimum required to be implemented in C and then
we can use these in higher level Acton based constructs, like we could
expose them via the OpenMetrics exposition format or something else.