Limit test perf concurrency to 1
While we have only run 1 concurrent test within a module when in perf
mode, multiple modules have been executed at the same time, so we've
gotten useless results. This fixes that by serializing the test modules
to run one by one.

We still run in full concurrency when not in perf mode.