Do not touch acton as part of building dist
I implemented this in order to have the acton binary in the dist reflect
the last modification of anything in the distribution. Since the acton
binary is included as a "source" file in the up-to-date check, if the
compiler was recompiled, we would also recompile Acton programs.

However, I realized that this is not necessary. We should only run the
actonc compilation of actually compiling .act files to .ty / .c / .h if
the compiler itself has changed. That is already reflected in its mtime.
Changes to C files in stdlib, RTS or builtins will be reflected in the
build output since the zig build is always run and we let zig figure out
if anything has changed there.