Fix str.strip() to handle empty strings
The str.strip() method was not handling empty strings correctly. We
would get back a \0 character instead of an empty string. Now fixed by
returning input string if length is 0. I don't think it's an issue that
you get back the actual string if you call strip on an empty string
rather than a new string with the same value as you'd otherwise get.