Fix re groups for no match
It is possible to define a matching group in a regexp that does not
match while the overall regexp does match, in which case the individual
group should be None. As such, the type has been updated to be a
list[?str] instead of list[str] for both group and named groups.

The code has been updated to avoid segfault since the string offsets are
unset (-1) when a group has no match.

New test case to test this.