Temp fix (CI) build via zigballs
Official zig build download page has removed the particular zig build we
used. We cannot directly go to zig v0.12 which is the closest release to
what we need, since some stuff we use was deprecated.

Meanwhile, we fix this by uploading the missing .tar.xz file to our own
little mirror of the zig builds and use that for our build.