Avoid deprecated anonymousDependency
Use the normal b.dependency() instead, just like all the other normal
Zig projects! However, it comes with some challenges, since
b.dependency("foo") refers to the dependency "foo" which maps up (as a
string) to the package defined in build.zig.zon with the name "foo".
Normally, the zig compiler takes the build.zig.zon and builds a
dependencies.zig that is then built together with build.zig /
build_runner.zig. However, zig does all sorts of stuffht at we don't
want, like it forces the use of absolute paths in the generated
dependencies.zig, based on paths that are relative in build.zig.zon.
This makes the final builder binary non-portable, so we got to avoid
that. Similarly, it uses content hashing or some sort of checksum to
address stuff, which is just sort of hard to work with as a human (I am
a human and I'm currently writing dependencies.zig manually). It'd
obviously be better if we could convince zig or write our own little
nifty program that reuses bits from zigs stdlib Build library to produce
the dependencies.zig, but I'm not familiar enough with this stuff to do
that. So in the meantime, here's chewing gum and tape.

Just as before, we inject our own dependencies.zig but it's rewritten so
we use relative paths that are relative in a project, specifically
$project/.build/sys should be a symlink that points to the Acton system
directory so we can access the dependencies in there. actonc now creates
.build/sys and sets it up as a symlink to the system dir. It already
knows about the system dir, since its own executable is within the
system dir.

For some reason I can't quite figure out, two dependencies (out of 10+),
libxml2 and mbedtls, tries to open something in their own dir but as a
dependency. I'm guessing this normally just works because the paths are
absolute, but now that they are relative, it tries to access like
.build/sys/deps/libxml2/.build/sys/deps/libxml2/include.. I have done
the worst hack imaginably and simply added a .build/sys symlink in
libxml2 and mbedtls that goes up to ../../../ so that these can be
resolved. Ugh *shudders*.

Anyhow, it works. One day zig will hopefully be better at supporting our
types of usecase and maybe then we can remove some of these hacks or we
can do a better job by writing something more specific in zig ourselves.