Support project dependencies
Well, we really did support it before but it broke with the recent zig
upgrade when we replaced the use of anonymousDependency with proper zig
b.dependency(). Our previous use for how we handled project dependencies
really was a hack and only ever could work for the most basic of cases.

Unlike before, we now generate and write a build.zig file to disk, so we
can customize it for our project, listing the actual dependencies it
has, which is exactly what we do. This also opens up for the
possibilities of customizations per project, for example to also link
with C libraries (that's not supported yet though).

It's a bit of a spaghetti going between acton and actonc as the duties
between them are not cleanly split. --keepbuild is not a pretty option
and we should strive to clean this up a bit. At the same time, we might
just want to accept a slightly messy situation and await zig to mature
as that might otherwise trigger more changes on our end anyway.