Support including Zig package dependencies
It is now possible to define a Zig package dependency in build.act.json,
our own variant of Zig's build.zig.zon. We use this information to
generate a build.zig.zong as well as inject in some places in build.zig,
in order for things to work correctly. And it does work! Very cool!

There's a slight change in behavior with regards to dependencies on
other Acton projects. They used to be discovered by being in the deps/
directory, like deps/foo is dependency package foo. Now they need to
also be defined in build.act.json. This is an improvement though as the
idea is that in the future we will offer automatic download of these and
then we can't discover them from a directory anyway, so doing config is
the right approach.

Generating build.zig and build.zig.zon is now done by acton rather than
actonc. This was the idea from the very beginning. It allows us to
gather a complete view in acton before proceeding to setup the build
environment and execute all compilation commands.