Do not pass --record when building tests
This is sort of a fine line to walk here. We currently use
build_cmd_args to take all arguments to acton and pass them through to
actonc. This works great since we don't have to manually specify all
arguments that should actually be passed through but on the other hand
it means that some arguments, which are not available on actonc, need to
be stripped out. In this case, --record is one such argument.