Stop building utf8proc for distribution
Just ship the source code and instead let zigbuilder build it when needed.

Also moves the dependency from the main build.zig to base's build.zig,
which we can do now that we are relying entirely on the Zig style build
and no manual linking of .a files, like we did before.

Bump version of utf8proc to one with a newer build.zig that installs the
header correctly and also has option for shared build (which we don't use).