Revamp zig build caching
Now split the cache into the zig global cache and the zig local cache.
Zig had these separate from the beginning but as Zig didn't
differentiate between them earlier, before the package manager, I didn't
know what they were good for and so we ended up just using a single
directory. Now that there is a package manager, we separate the global
cache, which is used for downloaded packages, and the local cache which
is used for built stuff.

The automatic clearing of the cache is kept but only applies for the
local cache. The global cache is never automatically cleared, instead a
warning is emitted once it crosses 10GB in size and for every additional
1GB in growth.