Move downloaded deps to .build/deps
We want them out of view.. I don't think we ever need to clear them.
This isn't a build cache like that, so it's fine for it to be
persistent. It would be ideal to move actonc output like .c / .h / .ty
files out-of-tree, but it's negligible compared to other things.

Also added cache size messages for global / local to be a little bit
more precise.