Guide: explain classes, protocols & dispatch
This started as an attempt at explaining the differences of method
dispatch for classes and protocols, but as it's inavoidable to mention
generics and that in turn touches types a bit, I also expanded some of
those sections. I have previously deliberately tried not to talk about
types very early on in the guide to avoid making it intimidating and we
strive to make it possible to write programs without having to
explicitly write loads of types.. but ... ehh, well, it's just hard to
talk about the language without saying a few words about types. Not
perfect, but I think this is an improvement.