Add Zig support in base for new base64 module
This adds a new stdlib module called base64. It does exactly what you
think it does. What's slightly cooler is the way it is implemented. Base
now includes a __root.zig file, which is the root module of base. It
defines two functions:

export fn base64Q_encode(data: *acton.str) callconv(.C) *acton.str {
export fn base64Q_decode(data: *acton.str) callconv(.C) *acton.str {

These are written in Zig but using the C ABI, 'export' saying they
should be exported, having the C calling convention and taking input and
output of the Zig type called `acton.str`. I have created this new acton
module in Zig which is meant to contain acton related things. In this
case some types, although so far I have only gotten to implement the
B_str type. Unlike C, Zig has modules so we don't have to treat
everything with a global name, that means acton.str can actually be
acton.str and doesn't collide with anything else called str in some
other module. Neat. Also, while Zig does expose a bunch of types that
are guarantee to match the C ABI, like c_int matches int in C on each
platform, I have opted to use more idiomatic Zig types since I think the
end goal is not to keep our C structs but to convert to Zig. Using more
Zig idiomatic types can save us some ugly intCasts. Obviously though,
the size of the ints have to line up, so I check that now with a
comptime test that inspects the type from Acton builtin (we import
builtin.h) and the Zig defined acton.str. Now, in this particular code
right now we still get an intCast, but this is because the C struct is
used a int and maybe it would be more natural with a word sized int
instead? I didn't want to change our builtins in this commit, so I left
that for now, but by changing the B_str nbytes / nchars to like u64, we
could get to use the very idiomatic usize in Zig which happens to map to
what the Zig base64 functions take and produce!

It's quite a lot of boiler plate, wrapping things up in C.. but it
works. Maybe we can find nicer patterns for how to do this or maybe
automatically generate this binding code from the compiler? It would be
pretty nice if we could just write a base64.ext.zig and in there have a

pub fn encode(data: *acton.str) *acton.str {

i.e. skip a lot of the C mumbo jumbo. I guess it's hard to avoid
creating our box.

Also added in rts/gc.zig, which is a Zig module for the BDW GC that we
are already using. I opted to inline this file instead of adding a
dependency on yet another thing.  While zig can fetch stuff, it doesn't
help us bundle that in the Acton distribution, so easier to inline.
Plus, I'm hoping we can test out this interface and try to upstream it.