Let local dep path take precedence over URL
If a dependency is available both remotely as well as locally via a
path, the local path should take precedence. A common scenario would be
that the dependency normally is available as a download from the
Internet but the user decides to fork the dependency and make changes to
it. After cloning the repo, we can use the dependency just by updating
the build.act.json and set the path to the local clone.