Guide: integrating a C library
This includes a page on how to integrate a C library into Acton code. We
use the zlib C library to implement an Acton zlib modue for compression
decompression. We go through memory allocation and other things one
needs to keep in mind when working with C code around Acton.