Guide: Bump up pkg stuff
No need for such a deep hierarchy, better to bump up. I think I felt
originally we would have hierarchies for acton packages and zig
packages, but like, from a story perspective, this just makes more sense
IMHO.

Also sneak in new security chapter. It's got nothing but it's a reminder
for me to write something, so in it goes.