Speed up CI run-linux by doing 1 iteration (not 50)
The iterations was meant at some point to catch errors but even with the
error in mind (we had a problem at the time) we never saw this testing
any errors, it's just too small and simple of a test. It's point and
thus I'm removing it. I think it should also speed up the test execution
by a bit since when we use the script style runacton execution, each
invocation actually gets its own file, which means it's compiled anew
and thus takes "some time". It feels in CI like this is 0.5~1s per
iteration, so we're looking at 20-60 seconds of time for a whole test!?