Allow naming overlap for hierarchical modules
Having a module foo (src/foo.act) and a foo.bar (src/foo/bar.act) would
previously "conflict" and lead to a compilation error. We now allow it. We
should really have some more checks around this to ensure that there are no
true conflicts, i.e. that the module foo does not define any "bar", since the
foo.bar is now a module... but that's for later ;) Meanwhile, one needs to be
careful :)