Fix hierarchical module import
We had some tricky cases where we got compilation errors due to imports
not working properly. It has to do with hierarchical modules, so like
when there is a foo.act and a foo/bar.act. If we import foo.bar first
into our compilation environment, the foo module will just be an empty
module, but it will exist and so if we later on import foo, the compiler
thinks it is already imported, won't read the .ty file and then proceed
to (likely at least) give some error about whatever we tried to access
in the foo module does not exist.

We now check explicitly if the module is empty. If it is, we try to
import it by reading .ty file. We cannot differentiate between an
actually empty module and a module that is just a placeholder, i.e. have
only been created to form the structure to a submodule, and so our only
choice is to attempt to read the .ty file whenever we have an empty
module. The caching we have here is useful but won't work for empty
module, however, re-reading empty modules likely isn't very expensive
and actually having empty modules is likely to be exceedingly rare in
real life.