Use BuildProject to compile files in project
If we run the compiler like `acton src/foo.act` in a project, we want to
still build the project like a project, which includes things like first
building dependencies. Previously, any invocation of acton on a single
file would invoke single file compilation regardless of the location of
the file. We now check if we are in a project and run the project build
but with the file specified, so only the specified file is compiled.