Proper class hierarchy for Dependency
We have to be able to separate the base type from the type for our pkg
dependencies, so renamed the latter to PkgDependency.

Previously we got an error when downloading zig deps since in the fetch
code we would check if it's a pkg dependency or zig dependency and then
delete from the corresponding list of outstanding downloads, but an
isinstance check on the pkg Dependency also matched the base type of
ZigDependency, and so we would try to delete it from the wrong dict and
get an error.