Add support for golden testing
We now support golden testing, i.e. comparing the output of a test with
an expected value, i.e. the golden value.

Golden values are saved in test/golden/MODULE_NAME/TEST_NAME

Errors, i.e. mismatches, are reported as NotEqualErrors, e.g.:

    kll@Boxy:~/dt/acton/test/stdlib_tests$ acton test --module test_testing3

    Tests - module test_testing3:
      foo:                   FAIL:  215 runs in 50.217ms
        testing.NotEqualError: Test output does not match expected golden value.
        Actual  : foobarBAR
        Expected: None

    1 out of 1 tests failed (0.410s)

And the expected output can be updated with --golden-update

We only support strings as golden values, although this could be
expanded in the future. Only unit tests and synchronous actor tests
support golden tests. This could also be expanded to support the
asynchronous and env tests.